Remove Publishable Model Mixin
A step on the way to turning models into ES classes. Gotta get rid of
them mixins.
